WIP: Detect the protocol of each side of the relay and do buffering if needed
If one end is TCP and the other end is websocket, then we need not send the
length prefix as WS already has length as part of its frame. So, we need to
chop the incoming TCP packets into just the records (without length prefix)
and pass it to WS write() or equiv function.

This commit only detects the protocol of each side and chops packets along
with length, so one websocket packet is one length-prefixed record. Next step
would be to use the modified client such that the length prefix can be stripped
off the record.